+++ src/options.rs
@@ -21,7 +21,7 @@ pub struct Config {
 
 #[derive(Default, Deserialize)]
 pub struct ConfigFlags {
-    pub avg_cpu: Option<bool>,
+    pub hide_average_cpu: Option<bool>,
     pub dot_marker: Option<bool>,
     pub temperature_type: Option<String>,
     pub rate: Option<u64>,
@@ -207,7 +207,7 @@ pub fn build_app(
     let app_config_fields = AppConfigFields {
         update_rate_in_milliseconds: get_update_rate_in_milliseconds(matches, config)?,
         temperature_type: get_temperature(matches, config)?,
-        show_average_cpu: get_avg_cpu(matches, config),
+        hide_average_cpu: get_show_average_cpu(matches, config),
         use_dot: get_use_dot(matches, config),
         left_legend: get_use_left_legend(matches, config),
         use_current_cpu_total: get_use_current_cpu_total(matches, config),
@@ -354,16 +354,19 @@ fn get_temperature(
     Ok(data_harvester::temperature::TemperatureType::Celsius)
 }
 
-fn get_avg_cpu(matches: &clap::ArgMatches<'static>, config: &Config) -> bool {
-    if matches.is_present("AVG_CPU") {
-        return true;
+/// Yes, this function gets whether to show average CPU (true) or not (false)
+fn get_show_average_cpu(matches: &clap::ArgMatches<'static>, config: &Config) -> bool {
+    // FIXME: Update the demo config file and default config files!  Need to remove
+    // old options and change to hide_avg_cpu option.
+    if matches.is_present("HIDE_AVG_CPU") {
+        return false;
     } else if let Some(flags) = &config.flags {
-        if let Some(avg_cpu) = flags.avg_cpu {
+        if let Some(avg_cpu) = flags.hide_average_cpu {
             return avg_cpu;
         }
     }
 
-    false
+    true
 }
 
 fn get_use_dot(matches: &clap::ArgMatches<'static>, config: &Config) -> bool {